requirements.txt
fastapi>=0.110.2,<0.111
uvicorn[standard]>=0.29,<0.30
pyyaml>=6.0.1,<7
